Refactor to extract conversion logic.
