+++ PDFToKeynote.xcworkspace/contents.xcworkspacedata
@@ -2,6 +2,9 @@
 <Workspace
    version = "1.0">
    <FileRef
+      location = "group:/Volumes/SSD/blue/Repos/PDFToKeynote/PDFToKeynote/Converter.swift">
+   </FileRef>
+   <FileRef
       location = "group:PDFToKeynote.xcodeproj">
    </FileRef>
    <FileRef
